added applied on to version info and change version to long

+++ src/MongoMigrator/IMigration.cs
@@ -3,6 +3,6 @@ namespace MongoMigrator
     public interface IMigration
     {
         void Up(IMongoDatabase mongoDatabase);
-        int GetVersion();
+        long GetVersion();
     }
 }
